Fixing a deadlock when transportDidClose called multiple times
The WebsocketsTransport could call the transportDidClose delegate method multiple times if the connection could not be open because the underlying WebSocket would first invoke the error handler and then the close handler and both of them called the transportDidClose. The easiest way to reproduce this was to try to start a connection to a SignalR app hosted on Azure where the webSocket support for the app service was disabled. This behavior was not correct. The fix is to introduce a flag to track if the transportDidClose delegate method was called and if so do not call it again.