Do not eagarly remove stream item handler on cancellation
When cancelling a stream invocation the client would immediately remove the
handler. This resulted in an error later when the client received a
completion message for the stream invocation. As a side effect the
stream invocation completion callback wouldn't be invoked. After the
change the handler will not be removed until the completion message was
received. Rather the client may receive stream items after cancellation
is processed by the server and would invoke the completion callback
after receiving the completion message. The previous logic also had a
bug where the client would stop processing stream items even if
the cancellation of a stream invocation failed.

Fixes: #66