Fix possible deadlock on stop
Trying to stop the connection before receiving any of the
open/error/close WebSocket event will cause a deadlock if stop was
called on the main thread. This is because by default SwiftWebSocket
events are invoked also on the main thread and `stop` will block this
thread making it impossible for the event handler to be invoked to
unblock `stop`. The fix is to instruct SwiftWebSocket to invoke event
handlers on a background thread.

Fixes: #91