Reconnection avalanche if negotiatation failed
If negotiation kept failing during reconnect there would be a never
ending reconnect avalanche that would never end. An additional side
effect would be that each time the connectionWillReconnect handler would
be invoked. The issue was that the internal counter of failed attempts
was not updated correctly and each reconnect was treated as the first
one. This was very easy to reproduce just by shutting down the server.