Check for watchOS to ignore allowSelfSignedSSL
This was the only change I needed to make to get this running under watchOS. Given I can test on an iOS device against a self signed cert end point, I have no reason to test the same thing on watchOS, so it's an easy fix for a very large gain. I also played around with #if os(iOS) however this might cut out tvOS, which I have never tried.