Fix deadlock caused by using an internal queue for callbacks
To fix a deadlock we started scheduling some callbacks to internal
queues. This can cause deadlocks in case the callback calls back into the
client. In this case a test was failing because it called `stop()` from
the `sendDidComplete` but `sendDidComplete` was invoked on the same
queue a the `stop` is trying to acquire to update its internal state.
For now changing the callbacks to the main queue but it should be
configurable via options.