Block setting skipNegotiation if webSocket transport is not available
Websocket task is not supported on some older OSX, iOS, watchOS and tvOS
versions and the client will fallback to longPolling. However it was
still possible to set the `skipNegotiation` to `true` which forced
creating a webSocket transport which failed because the webSocket
transport was not supported. This change blocks the ability to set
`skipNegotiation` if the webSocket transport is not supported.