Fix a regression that results in a deadlock
This refactoring f4558f659bead94ebac49988844cd3bc4703c8b1 effectively removed the fix
introduced in https://github.com/moozzyk/SignalR-Client-Swift/pull/244 and the client
started deadlocking if a keep alive was sent during reconnects. To avoid
deadlocks the callback will be now scheduled asynchronously on the callback queue if
there is an error due to a reconnect. For all other cases ReconnectableConnection will
rely on the fact that the underling HttpConnection already schedules callbacks to be
executed asynchronously.

Fixes: #264